backend/pyproject.toml
[tool.poetry]
name = "todo_list"
version = "0.1.0"
description = "Beejee todo list project"
authors = ["dmivek <dmitriyvek@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.1.3"
Flask-SQLAlchemy = "^2.5.1"
dynaconf = "^3.1.9"
Flask-Migrate = "^3.1.0"
psycopg2-binary = "^2.9.3"
graphene = ">2.0.1,<3.0.0"
Flask-GraphQL = "^2.0.1"
PyJWT = "^2.4.0"
loguru = "^0.6.0"
Flask-Cors = "^3.0.10"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
flake8 = "^5.0.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
